implement calls to new report viewers

+++ src/main/java/com/napier/sem/models/HighLevelPopulationReportModel.java
@@ -9,12 +9,25 @@ package com.napier.sem.models;
  */
 public class HighLevelPopulationReportModel {
 
+    /**
+     * The name of the geographic area
+     */
     public String Name;
 
-    public int Population;
+    /**
+     * The total population in the geographic area
+     */
+    public long Population;
 
-    public int CityPopulation;
+    /**
+     * The population of the cities in the geographic area
+     */
+    public long CityPopulation;
 
+    /**
+     * Calculates the percentage of the geographic area who reside in cities
+     * @return the percentage value as double
+     */
     public double GetPercentCityPopulation(){
         // return a zero to handle 0 population countries such as Antarctica
         if (Population == 0 ) return 0;
@@ -22,6 +35,10 @@ public class HighLevelPopulationReportModel {
         return ((double) CityPopulation / (double) Population) * 100;
     }
 
+    /**
+     * Calculates the percentage of the geographic area who do not reside in cities
+     * @return the percentage value as double
+     */
     public double GetPercentNotCityPopulation(){
         // return a zero to handle 0 population countries such as Antarctica
         if (Population == 0 ) return 0;
